Extracting to script module
